requirements.txt
wheel==0.37.0
lxml==4.6.4
requests==2.26.0
PyMySQL==1.0.2
PyPika==0.48.8
python-dotenv==0.19.1
tqdm==4.62.3
